Checks variable type before updating
This verifies that the variable type matches before updating.  Currently
it only checks the name, and then updates the value.  For instance, if
you have a machine variable, but accidently use set instead of
set_machine, the variables pane will show the updated value, but its not
actually updated because it created a player variable with that name in
the actual runtime.  This creates a new dictionary to tie the varaible
to the type, and then verifies the variable exists of the correct type
before updating a row.  Otherwise it will enter it.  In this case you
would have two variables named the same, one of type machine, and one of
type player.